fix: split useStatusBarHeightDetector hook into web and native flavours
